requirements.txt
django-filter==1.1.0
django==1.11
mysqlclient==1.3.12
djangorestframework==3.7.7
Markdown==2.6.11
Pillow==5.0.0
